Create AddStrings.py
Description
Given two non-negative integers num1 and num2 represented as string, return the sum of num1 and num2.

Note:

The length of both num1 and num2 is < 5100.

Both num1 and num2 contains only digits 0-9.

Both num1 and num2 does not contain any leading zero.

You MUST NOT convert the inputs to integer directly.

Complete the addStrings function. It takes the input as two strings & you need to return the string (which will same as the addition of two number as integers).