Create Grades.py
Suppose 10x Academy is planning a new grading system. It will be as follows:

Every student receives a grades in the inclusive range from 0 to 100.
Any grade less than 40 is a failing grade.
We also want to round each student's grade according to these rules:

If the difference between the grade and the next multiple of 5 is less than 3, round grade up to the next multiple of 5.
If the value of grade is less than 38, no rounding occurs as the result will still be a failing grade.
Examples:
grade = 84 round to 85 (85-84 is less than 3)
grade = 29 do not round (result is less than 40)
grade = 57 do not round (60-57 is 3 or higher)
Given the initial value of grade for each of n students, write code to automate the rounding process.

Input Format
The first line contains a single integer, n, the number of students. Each line i of the n subsequent lines contains a single integer, grades[i].

Output Format
Modified Grade of the students as one integer on each line.

Sample Input
4
73
67
38
33
Sample Output
75
67
40
33
Explanation
Student 1 received a 73, and the next multiple of 5 from 73 is 75. Since 75 - 73 < 3, the student's grade is rounded to 75.
Student 2 received a 67, and the next multiple of 5 from 67 is 70. Since 70 - 67 = 3, the grade will not be modified and the student's final grade is 67.
Student 3 received a 38, and the next multiple of 5 from 38 is 40. Since 40 - 38 < 3, the student's grade will be rounded to 40.
Student 4 received a grade below 38, so the grade will not be modified and the student's final grade is 33