Create Search Range.py
Given a sorted array of integers A of size N. Find the start and the ending index of the numbers in the given range (inclusive of start and ending point).

Your algorithm’s runtime complexity must be in the order of O(log n).

Return -1 -1 if there are no elements in the given range.

Input format
First line contains a positive integer N, denoting the number of elements in the array. The next line of the input contains N space seperated integers, which are the elements of the given array. The next line contains 2 integers, which specifies the range, we need to search.

Output format
2 space seperated integers, denoting the result.

Sample input1
6
1 2 3 3 4 5
3 4
Sample output1
2 4
Explanation1
The starting index of 3 is 2 and the ending index of 4 is 4.

Sample input2
10
1 2 3 3 4 4 7 7 10 10
5 6
Sample output2
-1 -1
Explanation2
There are no elements in the range of 5 to 6. Hence, the output is -1,-1.