Create Highlight.py
When a contiguous block of text is selected in a Document viewer, the selection is highlighted with a colored rectangle.

There is a list of 26 character heights aligned by index to their letters. For example, a is at index 0 and z is at index 25. There will also be a string. Using the letter heights given, determine the area of the rectangle highlight in mm2 assuming all letters are 1mm wide.

Input Format
The first line contains 26 space-separated integers describing the respective heights of each consecutive lowercase English letter, ascii[a-z].

The second line contains a single word consisting of lowercase English alphabetic letters.

Output Format
One integer denoting the area of the rectangle highlighted (in mm2)

Sample Input
1 3 1 3 1 4 1 3 2 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 5 7
zaba
Sample Output
28
Explanation
The tallest letter in zaba is z at 7mm. The selection area for this word is 4 x 1mm x 7mm = 28mm2.