Create Longest Word.py
You are giving an interview at on of your dream company. The interviewer asked you a question and the offer letter is just after this question. So the problem is as follows - :

Given a string S, find the length of the longest word in the string. The words are separated by Spaces. If no such word exists, print 0

Input
one line containing the String S

Output
one line containing the length of the longest word in the String S

Example
Input: hello world

Output: 5

Explanation
because both words are of length 5

Input: helloooooo world

Output:

10

Explanation
because "helloooooo" is of length 10