Create minAssociates.py
An Algorithm to find the no. of teams that can be created fulfilling the criteria
Input:
The i/p fn consists of five arguments.
num, An integer representing no. of associates.
skills, A list of integers representing the skill levels of associates.
minAssociates,an integer representing  minimum  no. of associates required for the team.
minLevel, an integer representing the lower limit of skill level inclusive
maxLevel, an integer representing the higher limit of skill level inclusive
Output:
Return an integer representing the total number of teams that can be formed per the criteria