Create Currency.py
The Russian Empire was, predominantly, a rural society spread over vast spaces. In 1913, 80% of the people were peasants. Soviet historiography proclaimed that the Russian Empire of the 19th century was characterized by systemic crisis, which impoverished the workers and peasants and culminated in the revolutions of the early 20th century. Recent research by Russian scholars discovered some currency coins from the excavation, which was being used during 19th century.

The uniqueness about these currency coins is that, it contains a unique code, which is a string of large length. Each character in the code string is a digit between 0 and 9. The value of the coin is determined by taking the sum of all its character-digits.

You need to find the value of the given coin, provided the unique code string.

Note: Try to solve it using recursion.

Input
The first line and the only line contains the unique code.

Output
A single line denoting the value of the given coin.

Example
Input:

789456123
Output:

45