Create checker.py
Description
The game of checker is very simple. You are given an array that has very interesting property. The property is a[i] > a[i+1] i.e array is sorted in reverse order . You are given another array and you need to search the contents of the given array in the first array. If the given number exists in the first array print its index in the first array else print -1.

Note:Consider 0 based indexing.

Input
First line is an integer T, representing total number of test cases. For each testcase, 1st line containing space separated distinct integers of array from left to right in which elements need to be searched. For each testcase, 2nd line contains number of elements to be searched. For each testcase, 3rd line containing space separated integers of array of element to be searched in given array.

Output
For each new line containing the result for each test input of a test case.

Example
Input:

2

9 8 7 6

2

6 5

6 2 1 0

3

1 2 5

Output:

3

-1

2

1

-1

Explanation
For the first test case, 9 8 7 6, 2 elements to be searched 6 is preset at position 3,5 is not present in array so print -1 For the second test case, 6 2 1 0, 3 elements to be searched 1 is present at position 2,1 is present at position 1,5 not present so print -1